backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.4.1
python-dotenv==1.0.0
google-cloud-texttospeech==2.33.0
google-auth==2.22.0
requests==2.31.0
google-generativeai==0.3.0
